ai_agent/requirements.txt
requests==2.31.0
python-dotenv==1.0.0
openai
pandas==2.1.1
numpy==1.24.3
dune-client
flask
tenacity